use smaller int type
We can save a couple of bytes.

+++ src/plugin/processor/voices.hpp
@@ -110,7 +110,7 @@ struct Voice {
     //
     u7 note_num = 0;
     MidiChannelNote midi_key_trigger = {};
-    int note_off_count = 0;
+    u16 note_off_count = 0;
 
     LFO lfo = {};
     OnePoleLowPassFilter lfo_smoother {}; // TODO(1.0): does the lfo need to be smoothed?
